+++ Energy.App.Standalone/Models/Tariffs/DefaultTariffDetail.cs
@@ -8,8 +8,8 @@ namespace Energy.App.Standalone.Models.Tariffs
 
         public DateTime DateAppliesFrom { get; init; }
         public MeterType MeterType { get; init; }
-        public double PencePerKWh { get; init; }
-        public double DailyStandingChargePence { get; init; }
+        public decimal PencePerKWh { get; init; }
+        public decimal DailyStandingChargePence { get; init; }
         public bool IsHourOfDayFixed { get; init; }
         public ICollection<DefaultHourOfDayPrice> DefaultHourOfDayPrices { get; init; }
